🚨 fix: ProfileDropdown - Dropdown으로 이름 변경, props를 받아서 유동적으로 활용가능하게 수정정
